fix(landing): split home page for SEO — remove duplicate schemas
